Added column count check if mapping provided
When loading a DBC file with a mapping you can now instantly verify, if
your column definition would - at least in amount of columns - match the
actual DBC file.

That does not mean the columns are correct themselves but just the
numbers match and would return data for each column.